ResponseGenerator에서 Response구조체 분리 후 리팩토링 #1900
