+++ .vscode/settings.json
@@ -76,6 +76,7 @@
     "concepts": "cpp",
     "__functional_03": "cpp",
     "memory_resource": "cpp",
-    "__functional_base_03": "cpp"
+    "__functional_base_03": "cpp",
+    "__memory": "cpp"
   }
 }
