+++ include/HttpProcessor.hpp
@@ -1,10 +1,13 @@
 #ifndef HttpProcessor_HPP
 #define HttpProcessor_HPP
 
+#include "CgiHandler.hpp"
 #include "Config.hpp"
+#include "Log.hpp"
 #include "MethodHandler.hpp"
 #include "PathFinder.hpp"
 #include "Request.hpp"
+#include "Response.hpp"
 #include "ResponseGenerator.hpp"
 
 class HttpProcessor
@@ -23,10 +26,9 @@ class HttpProcessor
  private:
   void setErrorPage(t_server server_data, Response& response_data);
   HttpProcessor(void);
-  
+
   // MethodHandler* m_method_handler; //이제 필요 없음?
   //   char* m_response_message;
-
 };
 
 #endif
