+++ src/HttpProcessor/PathFinder.cpp
@@ -1,6 +1,5 @@
 #include "PathFinder.hpp"
 
-#include "Log.hpp"
 // #include "./PathTest/PathFinder.hpp" // for test
 
 // #include <iostream>
