+++ src/ResponseGenerator/ResponseGenerator.cpp
@@ -5,73 +5,6 @@
 #include <iostream>
 #include <string>
 
-#include "Log.hpp"
-
-// canonical
-Response::Response()
-{
-  accepted_method = "";
-  rediretion_location = "";
-  file_path = "";
-  file_name = "";
-  root_path = "";
-  cgi_bin_path = "";
-  uploaded_path = "";
-  index_name = "";
-  index_exist = false;
-  status_code = OK_200;
-  redirection_exist = false;
-  auto_index = false;
-  file_exist = false;
-  path_exist = false;
-  cgi_flag = false;
-  is_cgi_timeout = false;
-  read_pipe_fd = -1;
-  write_pipe_fd = -1;
-  cgi_child_pid = -1;
-  error_keyword = false;
-  static_read_file_fd = -1;
-  static_write_file_fd = -1;
-}
-
-Response::Response(const Response& obj) { *this = obj; }
-
-Response::~Response() {}
-
-Response& Response::operator=(const Response& obj)
-{
-  if (this != &obj)
-  {
-    std::cout << "Response Body copy" << std::endl;
-    accepted_method = obj.accepted_method;
-    rediretion_location = obj.rediretion_location;
-    file_path = obj.file_path;
-    file_name = obj.file_name;
-    root_path = obj.root_path;
-    index_name = obj.index_name;
-    index_exist = obj.index_exist;
-    cgi_bin_path = obj.cgi_bin_path;
-    uploaded_path = obj.uploaded_path;
-    status_code = obj.status_code;
-    redirection_exist = obj.redirection_exist;
-    auto_index = obj.auto_index;
-    file_exist = obj.file_exist;
-    path_exist = obj.path_exist;
-    body = obj.body;
-    cgi_flag = obj.cgi_flag;
-    is_cgi_timeout = obj.is_cgi_timeout;
-    read_pipe_fd = obj.read_pipe_fd;
-    write_pipe_fd = obj.write_pipe_fd;
-    cgi_child_pid = obj.cgi_child_pid;
-    response_message = obj.response_message;
-    error_keyword = obj.error_keyword;
-    error_page_vector = obj.error_page_vector;
-    static_read_file_fd = obj.static_read_file_fd;
-    static_write_file_fd = obj.static_write_file_fd;
-  }
-  return (*this);
-}
-
 void ResponseGenerator::appendStrToResponse_message(std::string str)
 {
   m_response.response_message.insert(m_response.response_message.end(),
@@ -367,11 +300,16 @@ std::vector<char> ResponseGenerator::generateErrorResponseMessage()
   return (m_response.response_message);
 }
 
+// statuc code가 비정상이라면 내부적으로 generateErrorResponseMessage함수가
+// 호출되고 m_response.response_message를 생성 해줌
+// 1. cgi관련 request라면 1차 가공을 한 후
+// 2. start line생성
+// 3. header들 생성
+// 4. body 생성
 std::vector<char>& ResponseGenerator::generateResponseMessage()
 {
   try
   {
-    // LOG_DEBUG("m_response.status_code: %d", m_response.status_code);
     if (m_response.status_code != OK_200 && m_response.status_code != FOUND_302)
       throw(m_response.status_code);
     cgiDataProcess();
